+++ wtforglib/constants.py
@@ -1,3 +1,3 @@
 """Top-level module for wtforglib Library."""
 
-VERSION = "1.1.1"
+VERSION = "1.2.0"