+++ CHANGELOG.md
@@ -8,7 +8,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.1.1] - 2025-
+## [1.2.0] - 2025-04-06
+
+### Added
+
+- singleton class decorators
+
+### Changed
+
+- poetry >2.0 compatibility
+
+## [1.1.1] - 2025-03-03
 
 ### Added